In-progress double-sided popout;
